Rename mock files
