+++ src/__mocks__/node-info-mock.js
