+++ src/__mocks__/proposals-mock.js
