+++ src/__mocks__/rewards-mock.js
